合并至20240520 Lavac Hello World！
A=增加, B=修复, C=更改, D=删除

更新Readme
A:
	Minecraft 1.19.2 的服务器协议和3D旋转验证码
	Lavac:
		新增TypeResolver单独处理类型解析
		实现部分API
		加入了一键式的LavaCompiler，编译字符串形式的lambda
	FRP:
		使用Blake3替代不安全的SHA1哈希
		加入了nogui和autorun参数
	为FieldAccessor#getObject增加了基本类型自动包装
	基于ASM的基本类型数组Hasher
	NovelFrame支持循环替换
	DynByteBuf加入hex()和base64()
	基于ASM的事件总线，可继承，可取消注册，并且有错误提示
	加入Asn.1和Der编码的支持
	加入Jar完整性检查
	加入DateParser，方便的解析日期
	加入更高性能的TokenizerX实现便于未来替换Tokenizer
	加入了完美哈希表的实现
	加入了Autoload注解来自动加载类、Nixim或转换器
	加入了两种支持增量修改的分卷7z压缩格式
	加入了QOI图片格式的编码和解码
	QZArchiver现已支持尾部追加模式，在符合条件的情况下，可以减少硬盘写入并提高速度

B:
	定时任务可能意外的取消
	AnsiString#isBgColorRGB错误判断
	完成重构到一半的PacketBuffer
	Pipe本地连接CPU占用率高
	CCYaml和YAMLParser代码不同步
	Feedback CFB未遵循规范
	TextWriter性能很差
	BottomLine无法显示
	随机崩溃（懒常量解析）
	fastZip文件名处理错误（URIUtil.unescapeFileName）
	Mapper#loadCache未正确调用packup
	分卷压缩文件识别错误

C:
	ModuleKiller可以直接运行
	AnnotationRepo仅存储AccessData以降低内存占用
	Serializer: 不开启对象池的序列化会继续使用传统模式处理数组和列表
	FRP: 加入哈希白名单和自动登录功能
	重构日志记录器上下文
	HTTP: 重构定义式路由器，现在支持Vue-Router的路由语法
	修改插件加载方式，现在能在卸载时取消定时任务、注册的指令和HTTP路由
	重构AnnotationRepo和DeclFilter更易于使用
	整理了roj.collect中大部分的类
	将MutableInt合入CInteger
	将Nixim和ClassDefiner的调试代码移动至ClassDumper
	合并几个CLI工具到Archive4J
	CLI Console: Completion支持向前替换
	让Argument<File>更易于使用（加入了完整的tab补全支持）
	QZArchive的内存占用现已合并计算，避免构造特殊文件可能造成的意外内存溢出
	高精度计时器现在与keepalive daemon合并

D:
	删除了过时的Vinary编码
	删除了用不上的数学类

+++ java/roj/archive/ui/QZArchiver.java
@@ -11,7 +11,6 @@ import roj.concurrent.TaskPool;
 import roj.crypt.CRC32s;
 import roj.io.IOUtil;
 import roj.io.source.CacheSource;
-import roj.io.source.FileSource;
 import roj.io.source.FragmentSource;
 import roj.math.MathUtils;
 import roj.text.TextUtil;
@@ -40,7 +39,7 @@ import java.util.function.Consumer;
  * @since 2023/6/4 0004 3:54
  */
 public class QZArchiver {
-	public static final int AOP_REPLACE = 0, AOP_UPDATE = 1, AOP_UPDATE_EXISTING = 2, AOP_SYNC = 3, AOP_IGNORE = 4;
+	public static final int AOP_REPLACE = 0, AOP_UPDATE = 1, AOP_UPDATE_EXISTING = 2, AOP_SYNC = 3, AOP_IGNORE = 4, AOP_DIFF = 5;
 	public static final int PATH_RELATIVE = 0, PATH_FULL = 1, PATH_ABSOLUTE = 2;
 
 	public List<File> input;
@@ -59,7 +58,7 @@ public class QZArchiver {
 	public File cacheFolder;
 	public File outputFolder;
 	public String outputName;
-	public boolean singleThread;
+	public boolean singleThread, keepArchive;
 
 	public int autoSplitTaskSize;
 	public LZMA2Options autoSplitTaskOptions;
@@ -142,7 +141,7 @@ public class QZArchiver {
 		File out = new File(outputFolder, outputName.concat(splitSize == 0 ? "" : ".001"));
 		if (out.isFile()) {
 			try {
-				oldArchive = new QZArchive(out, password);
+				oldArchive = new QZArchive(ArchiveUtils.tryOpenSplitArchive(out, false), password);
 			} catch (Exception e) {
 				e.printStackTrace();
 			}
@@ -167,11 +166,7 @@ public class QZArchiver {
 				if (oldArchive != null) {
 					for (QZEntry entry : oldArchive.getEntriesByPresentOrder()) {
 						File file = byPath.get(entry.getName());
-						if (file == null || file.length() != entry.getSize() ||
-							(!fastAppendCheck ? contentSame(entry, file) :
-							(entry.hasModificationTime() && file.lastModified() <= entry.getModificationTime()) ||
-								(entry.hasCrc32() && checkCrc32(file, entry.getCrc32())))) {
-
+						if (file == null || isSame(file, entry)) {
 							byPath.remove(entry.getName());
 							keepEntry(entry);
 						}
@@ -194,11 +189,7 @@ public class QZArchiver {
 				if (oldArchive != null) {
 					for (QZEntry entry : oldArchive.getEntriesByPresentOrder()) {
 						File file = byPath.get(entry.getName());
-						if (file != null && file.length() == entry.getSize() &&
-							(!fastAppendCheck ? contentSame(entry, file) :
-							(entry.hasModificationTime() && file.lastModified() <= entry.getModificationTime()) ||
-								(entry.hasCrc32() && checkCrc32(file, entry.getCrc32())))) {
-
+						if (file != null && isSame(file, entry)) {
 							byPath.remove(entry.getName());
 							keepEntry(entry);
 						}
@@ -207,6 +198,22 @@ public class QZArchiver {
 			}
 			break;
 			case AOP_IGNORE: // 重新建立压缩包
+				break;
+			case AOP_DIFF: // 仅保留差异的文件
+				if (oldArchive != null) {
+					for (QZEntry entry : oldArchive.getEntriesByPresentOrder()) {
+						File file = byPath.get(entry.getName());
+						if (file == null) {
+							entry.setName("@@was_deleted/"+entry.getName());
+							keepEntry(entry);
+						} else if (isSame(file, entry)) {
+							byPath.remove(entry.getName());
+						} else {
+							entry.setName("@@was_changed/"+entry.getName());
+							keepEntry(entry);
+						}
+					}
+				}
 		}
 
 		// EMPTY
@@ -246,6 +253,13 @@ public class QZArchiver {
 		makeBlock(executable, chunkSize, coders, tmpa, tmpb);
 
 		return chunkSize;
+	}
+
+	private boolean isSame(File file, QZEntry entry) {
+		return file.length() == entry.getSize() &&
+			(!fastAppendCheck ? contentSame(entry, file) :
+				(entry.hasModificationTime() && file.lastModified() <= entry.getModificationTime()) ||
+					(entry.hasCrc32() && checkCrc32(file, entry.getCrc32())));
 	}
 
 	/**
@@ -406,23 +420,58 @@ public class QZArchiver {
 	}
 
 	public void compress(TaskPool pool, EasyProgressBar bar) throws IOException {
+		QZFileWriter writer;
 		File tmp;
-		do {
-			tmp = new File(outputFolder, outputName+"."+Integer.toString((int) System.nanoTime(),36)+".tmp");
-		} while (tmp.isFile());
 
 		long totalLength = 0;
 		for (WordBlockAppend block : appends) {
 			totalLength += block.size;
 		}
 
-		QZFileWriter writer;
-		if (splitSize == 0) {
-			IOUtil.allocSparseFile(tmp, totalLength);
-			writer = new QZFileWriter(new FileSource(tmp));
-		} else {
-			// .tmp.001
-			writer = new QZFileWriter(FragmentSource.fixed(tmp, splitSize));
+		createNewQZFW: {
+			if (keepSize > 0) {
+				long mySize = 0;
+				for (QZEntry entry : oldArchive.getEntriesByPresentOrder()) {
+					mySize += entry.getSize();
+				}
+
+				if (keepSize == mySize) {
+					System.out.println("FastPath: 尾部追加");
+
+					writer = oldArchive.append();
+
+					List<QZEntry> emptyFiles = keep.get(null);
+					if (emptyFiles != null) {
+						MyHashSet<QZEntry> set = new MyHashSet<>(emptyFiles);
+
+						SimpleList<QZEntry> files = writer.getEmptyFiles();
+						for (int i = files.size() - 1; i >= 0; i--) {
+							QZEntry file = files.get(i);
+							if (!set.contains(file)) {
+								writer.removeEmptyFile(i);
+							}
+						}
+					}
+
+					tmp = null;
+					keep.clear();
+					keepSize = 0;
+					oldArchive = null;
+					break createNewQZFW;
+				}
+			}
+
+			do {
+				tmp = new File(outputFolder, outputName+"."+Integer.toString((int) System.nanoTime()&Integer.MAX_VALUE,36)+".tmp");
+			} while (tmp.isFile());
+
+			if (splitSize == 0) {
+				IOUtil.allocSparseFile(tmp, totalLength);
+				writer = new QZFileWriter(tmp);
+			} else {
+				// .tmp.001
+				writer = new QZFileWriter(FragmentSource.fixed(tmp, splitSize));
+			}
 		}
 
 		if (bar != null) {
@@ -480,7 +529,8 @@ public class QZArchiver {
 
 		if (oldArchive != null) {
 			oldArchive.close();
-			Files.delete(new File(outputFolder, outputName).toPath());
+			if (!keepArchive)
+				Files.delete(new File(outputFolder, outputName).toPath());
 		}
 
 		if (bar != null) {
@@ -536,10 +586,12 @@ public class QZArchiver {
 			writer.setCompressHeader(0);
 		}
 
+		writer.flag |= QZWriter.TRIM_FILE;
 		writer.setIgnoreClose(false);
 		writer.close();
 		if (bar != null) bar.end("压缩成功");
-		Files.move(tmp.toPath(), new File(outputFolder, outputName).toPath());
+
+		if (tmp != null && !keepArchive) Files.move(tmp.toPath(), new File(outputFolder, outputName).toPath());
 	}
 
 	private QZWriter parallel(QZFileWriter qfw) throws IOException {
